requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
clarifai==2.6.1
Click==7.0
configparser==3.7.4
cryptography==2.6.1
decorator==4.4.0
defusedxml==0.6.0
emoji==0.5.2
entrypoints==0.3
ephem==3.7.6.0
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
future==0.17.1
googleapis-common-protos==1.6.0
grpcio==1.22.0
ibapi==9.76.1
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.0.0
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.11.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
Pygments==2.4.2
pyrsistent==0.15.2
PySocks==1.7.0
pytest==3.7.1
pytest-sugar==0.9.1
python-dateutil==2.8.0
python-telegram-bot==11.1.0
pyzmq==18.0.1
qtconsole==4.5.1
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.5
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
